add feature search for property and room

+++ src/main/java/com/daytoday/hostelrybooking/request/AddRoomRequest.java
@@ -15,10 +15,10 @@ public class AddRoomRequest {
     private Property property;
     private String roomName;
     private String description;
-    private int maxGuest;
+    private Short maxGuest;
     private BigDecimal pricePerNight;
     private String bedType;
-    private int unitAvailable;
+    private Short unitAvailable;
     private Double roomSize;
     private List<UUID> amenityIds;
 }
